Moved Commandline Parsing into own file
